Done refactoring.
