+++ components/spider-core/Cargo.toml
@@ -15,10 +15,12 @@ semver = "1.0.27"
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
 spider-derive = { path = "../spider-derive" }
+spider-utils = { path = "../spider-utils" }
 sqlx = { version = "0.8.6", features = ["mysql"] }
 strum = { version = "0.28.0", features = ["derive"] }
 thiserror = "2.0.18"
 
 [dev-dependencies]
+anyhow = "1.0.98"
 tokio = { version = "1.50.0", features = ["macros", "rt-multi-thread"] }
 tokio-util = { version = "0.7", features = ["rt"] }
