+++ components/spider-utils/Cargo.toml
@@ -8,6 +8,9 @@ name = "spider_utils"
 path = "src/lib.rs"
 
 [dependencies]
+rmp-serde = "1.3.1"
+serde = { version = "1.0.228", features = ["derive"] }
+thiserror = "2.0.18"
 tracing-appender = "0.2.5"
 tracing-subscriber = {
   version = "0.3.23",
