+++ components/spider-utils/src/lib.rs
@@ -1,3 +1,4 @@
 //! Shared utilities for Spider crates.
 
 pub mod logging;
+pub mod wire;
